server/requirements/requirements_api.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]>=3.3.0
requests>=2.31.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
